players part 11 and refactors

+++ pages/play/src/states/players/player.rs
@@ -1,11 +1,12 @@
 use std::sync::Arc;
 
-#[derive(Clone, Default)]
+#[derive(Clone, Default, PartialEq, Eq, Hash, serde::Deserialize)]
 pub struct Player {
     pub name: Option<Arc<str>>,
+    #[serde(rename = "systemName")]
     pub system: Option<Arc<str>>,
-    pub rank: u32,
+    pub rank: u8,
     pub account: bool,
     pub badge: Option<Arc<str>>,
-    pub medals: [u32; 5],
+    pub medals: [u8; 5],
 }
